+++ src/01_constants.sh
@@ -1,4 +1,4 @@
-SCRIPT_VERSION="0.3.1"
+SCRIPT_VERSION="0.3.2"
 INSTALL_PATH="/usr/local/bin/getout"
 UPDATE_URL="https://raw.githubusercontent.com/xiaochengshiguduo/getout/main/getout.sh"
 UPDATE_SHA256_URL="https://raw.githubusercontent.com/xiaochengshiguduo/getout/main/getout.sh.sha256"